Move ChecksumCalculator to Core (#1634)
