+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithStrongNames.txt
@@ -1,7 +1,7 @@
 ﻿---
 ################################
 # Assembly references report
-# Report date/time: 2020-08-18T17:15:37.7926624Z
+# Report date/time: 2020-08-18T19:50:10.8525256Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.10.0.217
@@ -19,7 +19,7 @@
 #
 ################################
 
-# Base directory: C:\Projects\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2019\net46
+# Base directory: C:\Projects\SLVS\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2019\net46
 Include patterns:
 - 'Sonar*'
 Exclude patterns:
@@ -134,6 +134,7 @@ Assembly: 'SonarLint.VisualStudio.Core, Version=4.26.0.0, Culture=neutral, Publi
 Relative path: 'SonarLint.VisualStudio.Core.dll'
 
 Referenced assemblies:
+- 'BouncyCastle.Crypto, Version=1.8.6.0, Culture=neutral, PublicKeyToken=0e99375e54769942'
 - 'mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed'
 - 'PresentationCore, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
@@ -144,14 +145,13 @@ Referenced assemblies:
 - 'System.IO.Abstractions, Version=7.0.0.0, Culture=neutral, PublicKeyToken=96bf224d23c43e59'
 - 'System.Net.Http, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'
 - 'System.Xml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 10
+# Number of references: 11
 
 ---
 Assembly: 'SonarLint.VisualStudio.Integration, Version=4.26.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 Relative path: 'SonarLint.VisualStudio.Integration.dll'
 
 Referenced assemblies:
-- 'BouncyCastle.Crypto, Version=1.8.6.0, Culture=neutral, PublicKeyToken=0e99375e54769942'
 - 'EnvDTE, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'
 - 'EnvDTE80, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'
 - 'Microsoft.Alm.Authentication, Version=4.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
@@ -182,7 +182,7 @@ Referenced assemblies:
 - 'System.Xml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.Xml.Linq, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'WindowsBase, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35'
-# Number of references: 31
+# Number of references: 30
 
 ---
 Assembly: 'SonarLint.VisualStudio.Integration.TeamExplorer, Version=4.26.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
